+++ js/ruta22.js
@@ -54,7 +54,6 @@
       '<p class="destacado">' + R.esc(actualidad.shortSummary) + '</p>' +
       R.firma(actualidad) +
       R.procedencia(actualidad) +
-      R.formatosGrupo(actualidad) +
       R.credencialesHilo(actualidad) +
       '<div class="article">' + R.articleBody(actualidad.articleBody) + '</div>' +
       '<div class="importa"><h3>Por qué importa acá</h3><p>' +